# week02/work/week02.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "week02",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/api/v1/students": {
      "post": {
        "summary": "创建学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": "学生对象"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "创建成功"
                }
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "summary": "根据姓名查询学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "学生姓名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student",
                  "description": "学生对象"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "summary": "根据id查询学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "学生id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student",
                  "description": "学生对象"
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "根据id更新学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "学生id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": "学生对象"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "更新成功"
                }
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "summary": "根据id删除学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "学生id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "删除成功"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/students/all": {
      "get": {
        "summary": "查询所有学生",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student",
                    "description": ""
                  },
                  "description": "所有学生列表"
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "Phone": {
        "type": "object",
        "properties": {
          "band": {
            "type": "string",
            "description": ""
          },
          "price": {
            "type": "number",
            "description": ""
          },
          "color": {
            "type": "string",
            "description": ""
          }
        }
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": ""
          },
          "gender": {
            "type": "string",
            "description": ""
          },
          "birthday": {
            "type": "string",
            "description": "日期"
          },
          "phone": {
            "$ref": "#/components/schemas/Phone",
            "description": ""
          }
        }
      }
    },
    "responses": {},
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}